Adding in some required utilties (#23)
* Adding in some required utilties

For the simulation next semester, it would be useful to let
the students have access to numpy as a math manipulation package.
I also need to have valgrind as we move to more use of docker
as our preferred way to run the grader. This should handle both
of the issues.

* Responding to PR

Good point. This should remove pip.