Install CPU-only Torch (#32)
We couldn't find much about the error Submitty was giving us online, but
ChatGPT said it had to do with trying to load CUDA which makes sense
since we don't have access to it, so I've changed the environment to
install the CPU-only version of torch. Hopefully, this fixes the error.